refactor: extract xtdb arrow preparation
