+++ polars_hist_db/backends/xtdb.py
@@ -27,6 +27,7 @@ from .base import (
     execute_table_health_query,
 )
 from .temporal import system_time_hint_clause
+from . import xtdb_arrow as _xtdb_arrow
 from . import xtdb_delta as _xtdb_delta
 from . import xtdb_query as _xtdb_query
 from . import xtdb_schema as _xtdb_schema
@@ -70,18 +71,21 @@ LOGGER = logging.getLogger(__name__)
 def __getattr__(name: str) -> Any:
     """Keep legacy private XTDB imports working during the module split."""
     try:
-        return getattr(_xtdb_query, name)
+        return getattr(_xtdb_arrow, name)
     except AttributeError:
         try:
-            return getattr(_xtdb_delta, name)
+            return getattr(_xtdb_query, name)
         except AttributeError:
             try:
-                return getattr(_xtdb_staging, name)
+                return getattr(_xtdb_delta, name)
             except AttributeError:
                 try:
-                    return getattr(_xtdb_schema, name)
+                    return getattr(_xtdb_staging, name)
                 except AttributeError:
-                    return getattr(_xtdb_transport, name)
+                    try:
+                        return getattr(_xtdb_schema, name)
+                    except AttributeError:
+                        return getattr(_xtdb_transport, name)
 
 
 def _load_flight_sql() -> Any:
